Restructure heat load profile generation; fix using peak load for EVs
