+++ doc/gb-model/system.rst
@@ -235,6 +235,9 @@ Baseline electricity demand is equivalent to all sources of electricity consumpt
 For instance, lighting, appliances, cooling.
 The profile for this load is the same as the profile for the load in the reference weather year, so that the impact of weather is represented (e.g. for cooling demand).
 
+By default, FES baseline electricity demand also includes resistive heating electricity demand.
+We extract this demand and include it instead in the :ref:`building heat electricity lead profile <system-load-heat>`.
+
 .. _system-load-heat:
 
 Building heat
@@ -242,13 +245,11 @@ Building heat
 
 Residential and Industry & Commercial (I&C) building heat demand is considered separately.
 Since we are only concerned with the electricity system, this load the electricity demand to operate technologies that generate heat.
-Importantly, this is **only building heat load met by heat pumps**.
-Direct electrical heating demand is considered within baseline electricity demand.
+Importantly, this is building heat load met by heat pumps **and** direct (resistive) heaters.
 
-
-For countries which have been reliant on heat pumps for many years already (e.g. Norway and Sweden), the reference weather year :ref:`system-load-baseline` profile would include the seasonal effect of heat pumps already.
+For countries which have been reliant on electrical for many years already (e.g. France, Norway and Sweden), the reference weather year :ref:`system-load-baseline` profile would include the seasonal effect of heat pumps already.
 To ensure we do not double-count this seasonal effect in both load profiles, we remove a simulated historical heat pump load from the historical baseline electricity load profile.
-This means that only this "building heat" load has that profile shape, whereas the baseline electricity profile's seasonality should now only reflect the use of direct electrical heating.
+This means that only this "building heat" load has that profile shape, whereas the baseline electricity profile should not reflect building heat seasonality at all.
 
 .. _system-load-dsr:
