+++ rules/gb-model.smk
@@ -35,8 +35,8 @@ rule compose_network:
     input:
         unpack(input_profile_tech),
         demands=expand(
-            resources("gb-model/{{fes_scenario}}/{demand_type}_demand/{{year}}.csv"),
-            demand_type=config["fes"]["gb"]["demand"]["Technology Detail"].keys(),
+            resources("gb-model/{{fes_scenario}}/{sector}_demand/{{year}}.csv"),
+            sector=["baseline_electricity", "iandc_heat", "ev", "residential_heat"],
         ),
         dsr=expand(
             resources("gb-model/{{fes_scenario}}/regional_{sector}_dsr_inc_eur.csv"),
