+++ scripts/gb_model/heat/heat_demand_electricity_load_profile.py
@@ -3,9 +3,9 @@
 # SPDX-License-Identifier: MIT
 
 """
-heat demand shape profile generator.
+heat demand electricity load profile generator.
 
-This script generates the heat demand shape for each sector.
+This script generates the heat demand electricity load profiles for each sector.
 """
 
 import logging
@@ -31,11 +31,11 @@ def process_demand_timeseries(
     year: int,
 ) -> pd.DataFrame:
     """
-    Generate load profile shape for heat pump demands in each sector.
+    Generate electricity load profile for all heat demand in each sector.
 
     Note that the energy totals are given in TWh and the heating mix in MWh.
-    However, since we are only interested in the shape of the resulting demand profile,
-    the units do not matter as they cancel out in the normalization step at the end.
+    However, since we normalise the energy totals after combining with heat pump COP, the mismatch doesn't matter.
+    The result is given in MWh.
 
     Args:
         demand_path (str): Filepath to the hourly heat load profiles generated by PyPSA-Eur
@@ -52,12 +52,14 @@ def process_demand_timeseries(
     cop_profile = (
         pd.read_csv(cop_profile_path, index_col=["time", "name"], parse_dates=["time"])
         .rename_axis(index=["snapshots", "node"], columns="technology")
+        .assign(resistive=1)
         .stack()
     )
     heating_mix = (
-        pd.read_csv(heating_mix_path, index_col=["technology", "year"])
+        pd.read_csv(heating_mix_path, index_col=["bus", "technology", "year"])
         .squeeze()
         .xs(year, level="year")
+        .rename_axis(["node", "technology"])
     )
 
     demand = (
@@ -79,8 +81,9 @@ def process_demand_timeseries(
     load = (
         demand_share.mul(energy_totals).dropna().filter(regex=f"{sector} (space|water)")
     )
+    demand_shape = load.div(cop_profile)
     electricity_demand = (
-        load.div(cop_profile)
+        demand_shape.div(demand_shape.groupby(["node", "technology"]).sum())
         .mul(heating_mix)
         .groupby(level=["snapshots", "node"])
         .sum()
@@ -98,11 +101,17 @@ def process_demand_timeseries(
             "Some nodes have NaN / infinite electrified heat demand; "
             "please check the input data and heating technology mix."
         )
-        electricity_demand = electricity_demand.replace([np.inf, -np.inf], 0).fillna(0)
-    # Normalize the load profile, as only its shape is actually meaningful
-    electricity_demand_shape = electricity_demand / electricity_demand.sum()
+    assert (
+        isclose := np.isclose(
+            electricity_demand.groupby("node").sum(), heating_mix.groupby("node").sum()
+        )
+    ).all(), (
+        "The total electrified heat demand does not match the heating technology mix for some nodes. "
+        "Please check the input data and heating technology mix. "
+        f"Nodes with mismatch: {electricity_demand[~isclose].index.tolist()}"
+    )
 
-    return electricity_demand_shape
+    return electricity_demand
 
 
 if __name__ == "__main__":
